Move PSTypeSynonymDeclaration to its own package
